Ensure language for group - artifact searches
If doing a lookup by group_id and artifact_id it can hapen that severl results come back. A librariy with the same coordines can exist in the java AND clojure name space. That's why a language parameter is passed now. If there is still no result for the requested langauge the fall back is without language.